+++ test/DecimalFloat.maximize.t.sol
@@ -10,14 +10,14 @@ contract DecimalFloatMaximizeTest is Test {
 
     /// 1 => 1e38
     function testMaximize1() external pure {
-        (int128 maximizedCoefficient, int128 maximizedExponent) = LibDecimalFloat.maximize(1, 0);
+        (int256 maximizedCoefficient, int256 maximizedExponent) = LibDecimalFloat.maximize(1, 0);
         assertEq(maximizedCoefficient, 1e38, "coefficient");
         assertEq(maximizedExponent, -38, "exponent");
     }
 
     /// 1e18 => 1e38
     function testMaximize1e18() external pure {
-        (int128 maximizedCoefficient, int128 maximizedExponent) = LibDecimalFloat.maximize(1e18, 0);
+        (int256 maximizedCoefficient, int256 maximizedExponent) = LibDecimalFloat.maximize(1e18, 0);
         assertEq(maximizedCoefficient, 1e38, "coefficient");
         assertEq(maximizedExponent, -20, "exponent");
     }
