+++ test/DecimalFloat.minimize.t.sol
@@ -10,14 +10,14 @@ contract DecimalFloatMinimizeTest is Test {
 
     /// 1 => 1
     function testMinimize1() external pure {
-        (int128 minimizedCoefficient, int128 minimizedExponent) = LibDecimalFloat.minimize(1, 0);
+        (int256 minimizedCoefficient, int256 minimizedExponent) = LibDecimalFloat.minimize(1, 0);
         assertEq(minimizedCoefficient, 1, "coefficient");
         assertEq(minimizedExponent, 0, "exponent");
     }
 
     /// 1e18 => 1e38
     function testMinimize1e18() external pure {
-        (int128 minimizedCoefficient, int128 minimizedExponent) = LibDecimalFloat.minimize(1e18, 0);
+        (int256 minimizedCoefficient, int256 minimizedExponent) = LibDecimalFloat.minimize(1e18, 0);
         assertEq(minimizedCoefficient, 1, "coefficient");
         assertEq(minimizedExponent, 18, "exponent");
     }
